API

GrapeMR.RF_pulse_analysis — Method

RF_pulse_analysis(cf::ControlField; attenuation_ref = 0.0, B1_ref = 1.0, power_ref = 500.0)

Calculate calibration analysis of shaped pulse

### Parameters:
- `cf::ControlField`: The control field object containing the RF waveform.
- `attenuation_ref::Float64`: Reference attenuation in dB (default is 0.0 dB).
- `B1_ref::Float64`: Reference RF field strength in Tesla (default is 1.0 T).
- `power_ref::Float64`: Reference RF power in Watts (default is 500.0 W).

### Returns:
A tuple of calculated values:
- `max_amp`: Maximum RF amplitude in Hertz (Hz).
- `max_amp_tesla`: Maximum RF amplitude in microtesla (µT).
- `attenuation_B1`: Attenuation in decibels (dB).
- `power_max_B1`: Maximum power in Watts (W).
- `power_average`: Average power in Watts (W).
- `pulse_energy`: Pulse energy in Joules (J).

GrapeMR.amplitudes_and_phases — Method

amplitudes_and_phases(cf::ControlField)

Calculates amplitudes and phases in Hz and radians

GrapeMR.bSSFP_RF — Method

bSSFP_RF(N::Int, nTR::Int, α::Real, TR::Float64)

Generates a bSSFP pulse sequence with bandwidth BW and flip angle flip_angle in radians

Arguments

N::Int: Points

nTR::Int: How many TRs

α::Real: Flip angle

`TR::Float64: Repetition time TR in seconds

Output

Nx1 array with bSSFP pulse amplitudes

Example

```julia bSSFP(gpoutput, folderpath="/path/to/folder")

If no path is provided, it saves the files inside the folder where the package was installed folder name format : yyyy-mm-dd

GrapeMR.bandwidth_factor — Method

bandwidth_factor(cf::ControlField)

Excitation frequency range

GrapeMR.bloch_matrix — Method

bloch_matrix(B1x::Float64, B1y::Float64, Bz::Float64, Γ1::Float64, Γ2::Float64)

bloch_matrix # Input B1x: (::Float64) - B1x step B1y: (::Float64) - B1x step Bz: (::Float64) - B1x step Γ1: (::Float64) - B1x step Γ2: (::Float64) - B1x step

# Output
- Calculated 4x4 Bloch matrix

GrapeMR.bohb_hyperopt — Method

hyperoptimization(spins::Vector{<:Spins}, 
                  gp::GrapeParams, 
                  Tc::LinRange, 
                  max_iter::Int; 
                  i::Int=5, 
                  poly_start::Vector{Float64}=[1e-1, 1e-2], 
                  poly_degree::Vector{Int}=[1, 2, 3], 
                  B1ref::Float64=1.0)

Performs hyperparameter optimization using Bayesian Optimization and Hyperband (BOHB) for selecting hyperparameters.

### Input
- `spins::Vector{<:Spins}`: A vector containing spin systems for the optimization process.
- `gp::GrapeParams`: Grape algorithm parameters, which include time points, cost function, and mask for which fields are being optimized.
- `Tc::LinRange`: Range for the time control points for spline interpolation.
- `max_iter::Int`: Maximum number of iterations for the optimizer.
- `i::Int=5`: Number of optimization evaluations.
- `poly_start::Vector{Float64}=[1e-1, 1e-2]`: Starting points for the polynomial learning rate.
- `poly_degree::Vector{Int}=[1, 2, 3]`: Degrees for the polynomial learning rate.
- `B1ref::Float64=1.0`: Reference B1 field amplitude for the RF pulse.

### Output
- An optimized object from the BOHB optimization process, containing results and optimization history.

GrapeMR.bruker_normalized_amplitudes_and_phases — Method

bruker_normalized_amplitudes_and_phases(cf::ControlField)

Calculates amplitudes and phases normalized to 100 and 180 deg for Bruker implementation on TopSpin. All negative phase values are added a 360deg phase

### Parameters:
- `cf::ControlField`: The control field object containing the RF waveform.

### Returns:
A tuple of normalized amplitudes and phases:
- `norm_amp`: Maximum RF amplitude in Hertz (Hz).
- `norm_phase`: Normalize phases in degrees.

GrapeMR.calculate_cost_gradient — Method

calculate_cost_gradient(cost_func::Num, dict::Dict, vars::Vector{Num})

Generic function that maps variables into the gradient calculation using the Symbolics package dM = Differential(vars[]) dCostM = expandderivatives(dM(costfunc)) P = Symbolics.unwrap(substitute.(dCostM, (dict,))[1])

Input

- 'cost_func::Num': Cost Function from pre-defined dictionary
- 'dict::Dict': Variable dictionary
- 'vars::Vector{Num}': Function variables, Mx, My, etc.

Output

- 'cost_gradient::Vector{Float64}':

GrapeMR.complex_signal — Method

complex_signal(cf::ControlField)

Calculates complex RF signal in Hz

GrapeMR.cost_function — Method

 cost_function(iso::Isochromat, cf::Symbol)

Function that gets called cost function form the cost function dictionary

Input

- 'iso::Isochromat': IRespective isochromat for the calculation
- 'cf::Symbol': Cost Function

Output

- '(cost_val, cost_gradient)::Tuple{Float64, Vector{Float64}}': Tuple with cost funciton value and 4x1 gradient.

GrapeMR.cost_offsets — Method

cost_offsets(control_field::ControlField, spin_system::Spin, offsets::Vector{Float64})

Calculates the cost values for a given control field (control_field) and spin system (spin_system) over a range of B0 inhomogeneities (offsets).

# Arguments
- `control_field::ControlField`: The control field object that contains the RF pulse sequence.
- `spin_system::Spin`: The spin system for which the cost function is to be calculated.
- `offsets::Vector{Float64}`: A vector of offset frequencies (in Hz) representing B0 inhomogeneities.

# Returns
- `cost_values::Vector{Float64}`: A vector containing the cost values corresponding to each offset in `B0_values`.

GrapeMR.countour_cost — Method

countour_cost(cost_matrix::Matrix{Float64})

Creates a contour plot of the cost function values over the range of B0 and B1 inhomogeneities.

# Arguments
- `cost_matrix::Matrix{Float64}`: A 2D matrix representing the cost values for combinations of B0 and B1 inhomogeneities.

# Returns
- `c::Plot`: A contour plot of the cost function map.

GrapeMR.create_cost_matrix — Method

create_cost_matrix(control_field::ControlField, spin_system::Spin, offsets::Vector{Float64}, b1_inhomogeneities::Vector{Float64})

Generates a cost matrix representing the cost function values across a range of B0 and B1 inhomogeneities.

# Arguments
- `control_field::ControlField`: The control field object that contains the RF pulse sequence
- `spin_system::Spin`: The spin system for which the cost function is to be calculated.
- `offsets::Vector{Float64}`: A vector of offset frequencies (in Hz) representing B0 inhomogeneities.
- `b1_inhomogeneities::Vector{Float64}`: A vector representing B1 inhomogeneity percentages.

# Returns
- `cost_matrix::Matrix{Float64}`: A 2D matrix where each element represents the cost value for a particular 
                                    combination of B0 and B1 values. The matrix is normalized by its maximum value.

GrapeMR.create_folder — Method

create_folder(path::String)

Creates a folder at the specified path if it doesn't already exist.

Arguments

path::String: The path where the folder will be created.

Output

Returns the path to the created folder. If the folder already exists, it simply returns the path.

GrapeMR.dynamics — Method

dynamics(cf::ControlField, spins::Spin)

Function that returns the Isochromat object with the already calculated dynamics.

# Input  
- cf::ControlField - Adjoint State
- spins::Spin

# Output
- iso::Isochromat

GrapeMR.forward_propagation — Method

forward_propagation # Input - cf: (::ControlField) - Control fields struct - s: (::Spins) - Spin struct

# Output
- Magnetization vector 4xN

GrapeMR.get_cost_and_gradient — Method

get_cost_and_gradient(iso::Isochromat, cost_expr::Num, vars::Vector{Num})

Generic function that maps variables into the gradient calculation using the Symbolics package dM = Differential(vars[1]) dCostM = expandderivatives(dM(costfunc)) P = Symbolics.unwrap(substitute.(dCostM, (dict,))[1])

Input

- 'iso::Isochromat': Isochromts
- 'cost_expr::Num': Cost Function
- 'vars::Vector{Num}': Function variables -> Mx, My, etc.

Output

- 'cost_func_val::Float64': Cost Function Value
- 'cost_gradient::Vector{Float64}':

GrapeMR.gradient — Method

gradient(χ::Matrix{Float64}, M::Matrix{Float64}, H::Matrix)

gradient # Input - χ = (::Matrix{Float64}) - Adjoint State - M = (::Matrix{Float64}) - Forward Propagation - H = (::Matrix) - Hamiltonian

# Output
- ΔJ - 1xN matrix

GrapeMR.grape — Method

grape(p::Parameters, cf::ControlField, spins::Vector{<:Spins})

Implements the grape algorithm [1]

### Input
- `op::OptimizationParams`: Parameters for the optimization itself: max iterations, 
- `gp::GrapeParams`: Parameters related to Grape itself: time points, cost function, mask for which fields are being optimized.
- `cf::ControlField`: Initial control field - spline function -
- `spins::Vector{<:Spins}`: Vector with all spins included in the optimization

### Output
A scruct cointaing all optimization results:
- `grape_output::GrapeOutput': Data type with the optimized control fields, spin information and spin dynamics.

GrapeMR.hard_RF — Method

hard_RF(N::Int, t_c::Float64)

Generates a cubic spline pulse

Arguments

N::Int: Points

t_c::Float64: Shaped pulse time in seconds

Output

1xN array with spline pulse amplitudes

Example

```julia bSSFP(gpoutput, folderpath="/path/to/folder")

If no path is provided, it saves the files inside the folder where the package was installed folder name format : yyyy-mm-dd

GrapeMR.heatmap_cost — Method

heatmap_cost(cost_matrix::Matrix{Float64}, offsets::Vector{Float64}, b1_inhomogeneities::Vector{Float64})

Generates a heatmap plot of the cost function values over a range of B0 and B1 inhomogeneities.

# Arguments
- `cost_matrix::Matrix{Float64}`: A 2D matrix representing the cost values for combinations of B0 and B1 inhomogeneities.
- `offsets::Vector{Float64}`: A vector of offset frequencies (in Hz) used to label the x-axis.
- `b1_inhomogeneities::Vector{Float64}`: A vector representing B1 inhomogeneity percentages used to label the y-axis.

# Returns
- `h::Plot`: A heatmap plot showing the cost function map with a color bar indicating cost values.

GrapeMR.integral_factor — Method

integral_factor(cf::ControlField)

Power or energy of the RF pulse integrated over its duration

GrapeMR.max_peak_amp — Method

max_peak_amp(B1::Vector{ComplexF64})

Returns the maximum RF amplitude in Hz

GrapeMR.max_peak_amp_tesla — Method

max_peak_amp_tesla(B1::Vector{ComplexF64})

Returns the maximum RF amplitude in Tesla

GrapeMR.plot_control_fields — Method

plot_control_fields(cf::ControlField; unit::String = "Hz")

Plot the control fields of an RF pulse in different units.

### Parameters:
- `cf::ControlField`: The control field object containing the RF waveform data (`B1x`, `B1y` components).
- `unit::String`: The unit for the plot. Supported units are:
    - `"Hz"`: Plots `B1x` and `B1y` in Hertz.
    - `"rad/s"`: Converts the control fields to rads/sec and plots the amplitude and phase.
    - `"Tesla"`: Converts the control fields to Tesla using the gyromagnetic ratio for ¹H and plots in microtesla (µT).

### Returns:
- A plot object displaying the control fields in the specified units with amplitude and phase information.

GrapeMR.plot_cost_offset — Method

plot_cost_offset(cost_profile::Vector{Float64}, B0_values::Vector{Float64})

Creates a plot of the cost values as a function of B0 offset frequencies.

# Arguments
- `cost_profile::Vector{Float64}`: A vector containing the cost values to be plotted.
- `B0_values::Vector{Float64}`: A vector of offset frequencies (in Hz) corresponding to the cost values.

# Returns
- `p::Plot`: A plot object displaying the cost function offset profile.

GrapeMR.plot_cost_values — Method

plot_cost_values(cost::Vector{Float64}, gp::GrapeParams)

Plot the cost function convergence over iterations during the GRAPE optimization process.

### Parameters:
- `cost::Vector{Float64}`: A vector containing the cost values at each iteration.
- `gp::GrapeParams`: The parameters of the GRAPE optimization, including the cost function name.

### Returns:
- A plot object displaying the convergence of the cost function over the iterations.

GrapeMR.plot_magnetization_control_field — Method

plot_magnetization_control_field(cf::ControlField, isos::Vector{Isochromat})

Plot the magnetization trajectory for a set of isochromats and the corresponding control field.

### Parameters:
- `cf::ControlField`: The control field object containing the RF waveform data (`B1x`, `B1y` components).
- `isos::Vector{Isochromat}`: A vector of isochromat objects representing different spin systems.

### Returns:
- A plot object displaying:
    1. The magnetization trajectory in the transverse plane.
    2. The control field amplitude and phase over time.

GrapeMR.plot_transverse_magnetization — Method

2D Magnetization Plots

GrapeMR.random_sampler — Method

random_sampler(spins::Vector{<:Spins}, 
              gp::GrapeParams, 
              Tc::LinRange, 
              max_iter::StepRange; 
              i::Int=50, 
              poly_start::Vector{Float64}=[1e-1, 1e-2], 
              poly_degree::Vector{Int}=[1, 2, 3], 
              B1ref::Float64=1.0)

Performs random sampling for hyperparameter optimization.

### Input
- `spins::Vector{<:Spins}`: A vector containing spin systems for the optimization process.
- `gp::GrapeParams`: Grape algorithm parameters, which include time points, cost function, and mask for which fields are being optimized.
- `Tc::LinRange`: Range for the time control points for spline interpolation.
- `max_iter::StepRange`: Range for the maximum number of iterations for the optimizer.
- `i::Int=50`: Number of random samples to evaluate.
- `poly_start::Vector{Float64}=[1e-1, 1e-2]`: Starting points for the polynomial learning rate.
- `poly_degree::Vector{Int}=[1, 2, 3]`: Degrees for the polynomial learning rate.
- `B1ref::Float64=1.0`: Reference B1 field amplitude for the RF pulse.

### Output
- A random hyperparameter optimization object using random sampling.

GrapeMR.run_cost_analysis — Method

run_cost_analysis(grape_output::GrapeOutput, offset::Float64, b1_inhomogeneity_percent::Int)

Runs a complete cost analysis for the specified GrapeOutput object over a range of B0 and B1 inhomogeneities, and generates plots to visualize the cost function.

# Arguments
- `control_field::ControlField`: The control field object that contains the RF pulse sequence.
- `spin_system::Spin`: The spin system for which the cost function is to be calculated.
- `offset::Float64`: The range of offset frequencies (in Hz) for the B0 inhomogeneities.
- `b1_inhomogeneity_percent::Int`: The percentage of B1 inhomogeneity to consider for the analysis.

# Workflow
1. Generates a range of B0 values from `-offset` to `offset`.
2. Computes cost values for the specified B0 and B1 inhomogeneities using `cost_offsets` and `create_cost_matrix`.
3. Plots the cost function offset profile using `plot_cost_offset`.
4. Generates a heatmap of the cost matrix using `heatmap_cost`.
5. Creates a contour plot of the cost matrix using `countour_cost`.

# Example
run_cost_analysis(grape_output, 50.0, 30)

GrapeMR.save_bohb_data — Method

save_bohb_data(bohb::GrapeMR.GrapeOutput; folder_path = pwd())

Save data related to BOHB optimization into a folder organized by date.

Arguments

bohb::GrapeMR.GrapeOutput: The output from a BOHB optimization process.

folder_path::String = pwd(): The folder path where data will be saved. Defaults to the current working directory.

Output

Returns the full path to the folder where the data was saved.

Saved Files

bohb.jld2: Contains BOHB result in JLD2 format.

GrapeMR.save_grape_data — Method

save_grape_data(go::GrapeMR.GrapeOutput; folder_path = pwd())

Save data related to GRAPE optimization into a folder organized by date.

Arguments

go::GrapeMR.GrapeOutput: The output from a GRAPE optimization process.

folder_path::String = pwd(): The folder path where data will be saved. Defaults to the current working directory.

Output

Returns the full path to the folder where the data was saved.

Saved Files

grape_output.jld2: Contains the entire GrapeOutput struct in JLD2 format.

dict_cost_values.csv: CSV file containing the cost values from the optimization process.

dict_control_field.csv: CSV file containing control field values (B1x, B1y, Bz, and RF time).

dict_iso_spins.csv: CSV file containing isochromat spin parameters.

If no path is provided, it saves the files inside the folder where the package was installed folder name format : yyyy-mm-dd

GrapeMR.sinc_RF — Method

sinc_RF(N::Int, t_c::Float64, BW_Hz::Real, flip_angle::Float64)

Generates a sinc pulse with bandwidth BW and flip angle flip_angle in radians

Arguments

N::Int: Points

t_c::Float64: Shaped pulse time in seconds

BW_Hz::Real: Pulse bandwidth

flip_angle::Float64: Flip angle

Output

1xN array with sinc pulse amplitudes

Example

```julia bSSFP(gpoutput, folderpath="/path/to/folder")

If no path is provided, it saves the files inside the folder where the package was installed folder name format : yyyy-mm-dd

GrapeMR.spline_RF — Method

spline_RF(N::Int, t_c::Float64)

Generates a cubic spline pulse

Arguments

N::Int: Points

t_c::Float64: Shaped pulse time in seconds

B1_ref::Float64: Reference, normally pulse maximum amplitude

Output

ControlField struct

Example

```julia bSSFP(gpoutput, folderpath="/path/to/folder")

If no path is provided, it saves the files inside the folder where the package was installed folder name format : yyyy-mm-dd

GrapeMR.steady_state — Method

steady_state(s::GrapeMR.SteadyState)

Calculates the steady-state signal of a bSSFP sequence for a given set of parameters. The function calculates the magnetization evolution using repeated RF excitations and free precession steps until the steady state is reached.

# Arguments
- `s::GrapeMR.SteadyState`: A struct containing TR (Repetition Time), TE (Echo Time), T1 (Spin-lattice relaxation time), 
                            T2 (Spin-spin relaxation time), α (Flip angle in radians), Δϕ (Phase cycling increment), 
                            M_init (Initial magnetization vector), and B0inho (B0 inhomogeneity).
# Returns
- `signal::Vector{ComplexF64}`: The complex steady-state signal of the spin system.

GrapeMR.steady_state_geometric — Method

steady_state_geometric(s::GrapeMR.SteadyState)

Calculates the transverse steady-state magnetization using a geometric solution for a bSSFP sequence. Uses a geometric approach to derive the transverse component of the steady-state magnetization.

# Arguments
- `s::GrapeMR.SteadyState`: Struct containing the sequence parameters and spin system properties.

# Returns
- `Mxy::Vector{Float64}`: The steady-state transverse magnetization magnitude `Mxy`

GrapeMR.steady_state_geometric_Mz — Method

steady_state_geometric_Mz(s::GrapeMR.SteadyState)

Calculates the longitudinal steady-state magnetization using a geometric solution for a bSSFP sequence. Uses a geometric approach to derive the longitudinal component of the steady-state magnetization.

# Arguments
- `s::GrapeMR.SteadyState`: Struct containing the sequence parameters and spin system properties.

# Returns
- `Mz::Vector{Float64}`: The steady-state longitudinal magnetization magnitude `Mz`

GrapeMR.update! — Method

update(cf::ControlField, ∇xy::Tuple, ϵ::Float64)

update # Input - cf: (::ControlField) - Control fields struct - ∇xy: (::Tuple) - Calculated gradients for x and y components - ϵ: (::Float64) - Weigth of gradient

# Output
- Control Field - 1xN matrix